+++ src/main/java/com/dairy/service/FeedTypeService.java
@@ -12,7 +12,7 @@ public interface FeedTypeService {
 
 	public List<FeedTypeResponseDto> getAllFeedTypes();
 
-	public List<FeedTypeResponseDto> getFeedTypeByFeedCompanyId(int id, int branchId);
+	public List<FeedTypeResponseDto> getFeedTypeByFeedCompanyId(long id, int branchId);
 
 	public FeedTypeResponseDto findById(long id);
